Created a productBatch template, and modified the productBatchController by assigning get and post mappings for "/productBatch". Product Batch template currently only has newProductBatch form. Added a button on portalLanding for the productBatch template.

+++ src/main/java/com/storagemanager/storagemanager/productBatch/ProductBatchRepository.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface ProductBatchRepository extends CrudRepository<ProductBatchEntry, Long> {
 
     List<ProductBatchEntry> findAll();
-    List<ProductBatchEntry> findByproductSku(long productSku);
+    List<ProductBatchEntry> findByproductSku(int productSku);
 }
